requirements.txt
langchain>=0.2
langchain-community>=0.2
langchain-groq>=0.1.7
chromadb>=0.5.4
sentence-transformers>=2.5
gradio>=4.44
pydantic>=2
python-dotenv>=1.0
fastapi>=0.111
uvicorn>=0.30
